break out QuotedStatus from composer
