+++ app/javascript/mastodon/components/status/styles.module.scss
@@ -102,24 +102,6 @@
   z-index: 1;
 }
 
-// Overrides Card .body a:
-.quoteBody article {
-  a {
-    color: inherit;
-  }
-}
-
-.quoteError {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.quoteInfoPopup {
-  padding: var(--space-sm);
-}
-
 .meta {
   @include mixins.type-micro;
